fix: prevent rewriting import => import type in .svelte
Biome support for Svelte is currently still in the works. The tool can
only parse the <script> tags in .svelte files, which means Biome doesn't
know if a non-type import is used in the HTML section, treating it as
just a type by default if not directly used in the rest of the script
section. This breaks the app, so for now we turn off checking this
linting rule in .svelte files.