feat: editable table in detail view for selected nodes
This requires the svelte-headless-table dependency to build a table
synced with a svelte store. We create a custom cell component that is
editable with a small form to allow editing values. We also generate the
columns and groups based on arbitrary nested data on each node. We use
the "flat" and "remeda" dependencies so that we don't have to implement
very common functionality from scratch in JavaScript. Oddly enough,
capitalization is missing...